+++ lib/core/device/session/pexpect_wrapper/output_buffer.py
@@ -4,6 +4,8 @@ import regex
 
 from lib.services import logger
 
+from .common import clean_by_pattern
+
 NOFLAG = 0
 
 
@@ -79,8 +81,9 @@ def _split_flag_and_pattern(pattern):
 
 
 class OutputBuffer:
-    def __init__(self):
+    def __init__(self, clean_patterns=None):
         self.output = ""
+        self.clean_patterns = clean_patterns or {}
 
     def __str__(self):
         return self.output
@@ -96,7 +99,8 @@ class OutputBuffer:
         return len(self.output)
 
     def append(self, output):
-        self.output += output
+        original_output = self.output + output
+        self.output = clean_by_pattern(original_output, self.clean_patterns)
 
     def clear(self, pos=None):
         self.output = "" if pos is None else self.output[pos:]